download cairo and pixman from dl.feenk.com

+++ src/components/location.rs
@@ -13,6 +13,7 @@ pub enum LibraryLocation {
     Git(GitLocation),
     Path(PathLocation),
     Tar(TarUrlLocation),
+    Zip(ZipUrlLocation),
     Multiple(Vec<LibraryLocation>),
 }
 
@@ -36,7 +37,12 @@ impl LibraryLocation {
 
                 Ok(())
             }
-            LibraryLocation::Tar(tar) => tar.ensure_sources(default_source_directory, context),
+            LibraryLocation::Tar(tar_location) => {
+                tar_location.ensure_sources(default_source_directory, context)
+            }
+            LibraryLocation::Zip(zip_location) => {
+                zip_location.ensure_sources(default_source_directory, context)
+            }
         }
     }
 }
@@ -283,6 +289,72 @@ impl TarUrlLocation {
                     archive.unpack(&source_directory)?;
                 }
             }
+
+            std::fs::remove_file(&downloaded_path)?;
+
+            if let Some(ref sources) = self.sources {
+                let copy_options = fs_extra::dir::CopyOptions {
+                    content_only: true,
+                    ..Default::default()
+                };
+
+                fs_extra::dir::copy(
+                    source_directory.join(sources),
+                    &source_directory,
+                    &copy_options,
+                )?;
+
+                std::fs::remove_dir_all(source_directory.join(sources))?;
+            }
+        }
+        Ok(())
+    }
+}
+
+#[derive(Debug, Clone)]
+pub struct ZipUrlLocation {
+    url: String,
+    sources: Option<PathBuf>,
+}
+
+impl ZipUrlLocation {
+    pub fn new(url: impl Into<String>) -> Self {
+        Self {
+            url: url.into(),
+            sources: None,
+        }
+    }
+
+    pub fn sources(self, folder: impl Into<PathBuf>) -> Self {
+        Self {
+            url: self.url,
+            sources: Some(folder.into()),
+        }
+    }
+
+    fn ensure_sources(
+        &self,
+        default_source_directory: &Path,
+        context: &LibraryCompilationContext,
+    ) -> Result<(), Box<dyn Error>> {
+        let source_directory = context.sources_root().join(default_source_directory);
+
+        if !source_directory.exists() {
+            std::fs::create_dir_all(&source_directory)?;
+
+            let mut downloader = Downloader::builder()
+                .download_folder(&source_directory)
+                .build()?;
+
+            let to_download = Download::new(&self.url);
+
+            let mut result = downloader.download(&[to_download])?;
+            let download_result = result.remove(0)?;
+            let downloaded_path = download_result.file_name;
+
+            let downloaded_zip = File::open(&downloaded_path)?;
+            let mut archive = zip::ZipArchive::new(downloaded_zip)?;
+            archive.extract(&source_directory)?;
 
             std::fs::remove_file(&downloaded_path)?;
